fix: replace error arrays with free-form objects
Taikun's Swagger specification does not specify all the possible return
types of a Bad Request (400) error. It only specifies one. The ability
to specify multiple response types in Swagger, with the keyword `oneOf`,
came with versions 3.X only, we are using version 2.0.  The only way for
the generated Go client to be able to parse the response when it doesn't
match the specified model is to replace this model in the Swagger
specification by a free-type object which will match whatever the API
response is.

Replaces

  "type": "array",
  "items": {
    "$ref": "#/definitions/Error"
  }

With

  "type": "object"